# go-backend/internal/apidocs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JustRAG Public API",
    "description": "API for interacting with JustRAG knowledge bases. Authenticate using API keys generated from your user profile.\n\n## Authentication\nAll endpoints require a Bearer token (API key with `jrag_` prefix) in the `Authorization` header.\n\n## Streaming\nChat endpoints support Server-Sent Events (SSE) streaming by default. Set `?stream=false` to receive a synchronous JSON response instead.\n\n### SSE Format\n```\ndata: {\"type\": \"...\", ...}\n\ndata: [DONE]\n```",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Public API"
    },
    {
      "url": "/openai/v1",
      "description": "OpenAI Compatible"
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/models": {
      "get": {
        "tags": ["OpenAI Compatible"],
        "summary": "List models",
        "description": "Lists all accessible knowledge bases as OpenAI-compatible models. Each knowledge base is exposed as a model with ID in the format `kb-<uuid>`.",
        "operationId": "listModels",
        "servers": [{ "url": "/openai/v1" }],
        "responses": {
          "200": {
            "description": "List of models",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "object": { "type": "string", "enum": ["list"] },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Model" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OpenAIError" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OpenAIError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key with `jrag_` prefix. Pass as `Authorization: Bearer jrag_...`"
      }
    },
    "schemas": {
...
      "Model": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "kb-550e8400-e29b-41d4-a716-446655440000" },
          "object": { "type": "string", "enum": ["model"] },
          "created": { "type": "integer" },
          "owned_by": { "type": "string", "enum": ["justrag"] },
          "name": { "type": "string" }
        }
      },
...
      "OpenAIError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "type": { "type": "string", "enum": ["invalid_request_error"] },
              "code": { "type": "string", "enum": ["invalid_request"] }
            }
          }
        }
      }
    },
...
  }
}